feat: Retrieve hex data from data.json for applying effects
Previously, the hex data to be sent was stored within the Python file for convenience. With this change, we now retrieve the hex data from the data.json file, allowing us to apply any effect stored in the JSON by defining the effect name. This enhances flexibility and ease of use.

Changes Made:
- Updated hex data retrieval mechanism to read from data.json
- Improved flexibility in applying effects by defining effect name